Split swarms in own files
