+++ zeronet-swarm/lib/zero/nat/index.js
